+++ src/main/java/com/example/comprehensivedegisn/api_client/open_api/OpenApiClient.java
@@ -2,7 +2,7 @@ package com.example.comprehensivedegisn.api_client.open_api;
 
 
 import com.example.comprehensivedegisn.api_client.ApiClient;
-import com.example.comprehensivedegisn.api_client.dto.OpenApiRequest;
+import com.example.comprehensivedegisn.dto.request.OpenApiRequest;
 import com.example.comprehensivedegisn.batch.open_api.OpenAPiProperties;
 import com.example.comprehensivedegisn.batch.open_api.dto.ApartmentDetailResponse;
 import lombok.RequiredArgsConstructor;
